Remove `iap` packages and move models to `common`
