+++ app/src/main/java/com/trx/consumer/models/common/PurchaseModel.kt
@@ -1,4 +1,4 @@
-package com.trx.consumer.models.common.iap
+package com.trx.consumer.models.common
 
 import org.json.JSONObject
