+++ app/src/main/java/com/trx/consumer/screens/subscriptions/SubscriptionsFragment.kt
@@ -14,7 +14,7 @@ import com.trx.consumer.managers.LogManager
 import com.trx.consumer.managers.NavigationManager
 import com.trx.consumer.managers.UtilityManager
 import com.trx.consumer.models.common.AlertModel
-import com.trx.consumer.models.common.iap.SubscriptionModel
+import com.trx.consumer.models.common.SubscriptionModel
 import com.trx.consumer.screens.alert.AlertViewState
 import com.trx.consumer.screens.erroralert.ErrorAlertModel
 import com.trx.consumer.screens.subscriptions.list.SubscriptionsAdapter
