+++ app/src/main/java/com/trx/consumer/screens/subscriptions/list/SubscriptionsAdapter.kt
@@ -6,7 +6,7 @@ import com.trx.consumer.R
 import com.trx.consumer.common.CommonRecyclerViewAdapter
 import com.trx.consumer.common.CommonViewHolder
 import com.trx.consumer.managers.LogManager
-import com.trx.consumer.models.common.iap.SubscriptionModel
+import com.trx.consumer.models.common.SubscriptionModel
 import com.trx.consumer.views.EmptyViewHolder
 import kotlinx.coroutines.CoroutineScope
