+++ app/src/main/java/com/trx/consumer/screens/subscriptions/list/SubscriptionsViewHolder.kt
@@ -5,7 +5,7 @@ import com.trx.consumer.R
 import com.trx.consumer.common.CommonButton
 import com.trx.consumer.common.CommonLabel
 import com.trx.consumer.common.CommonViewHolder
-import com.trx.consumer.models.common.iap.SubscriptionModel
+import com.trx.consumer.models.common.SubscriptionModel
 
 class SubscriptionsViewHolder(view: View) : CommonViewHolder(view) {
